image Fade JQuery
**JQuery script setup for Hype 3.0 and Hype Professional**


Step 1. You need to access the document head by clicking on the document icon which will reveal Hype's Document pallette.

Step 2. You need to add the JQuery location for the HTML document. In this case we will use the version stored on Google's servers.

```
<script src="http://ajax.googleapis.com/ajax/libs/jquery/1.8.3/jquery.min.js"></script>

```

Step 3. You need to create a function that works with the JQuery Syntax that you require for your effect.
For example when a user clicks on an image the following will make the image fade out.
	
	//element - DOMHTMLElement that triggered this function being called
	//event - event that triggered this function being called
	function fadeout(hypeDocument, element, event) {
	$("#photoheadshot").fadeOut();
	}